app/requirements.txt
fastapi
uvicorn[standard]
databricks-sdk
openai
psycopg2-binary
